+++ lang/Cargo.toml
@@ -14,6 +14,8 @@ solana-program = "2.0.3"
 spl-token-2022 = { version = "4.0.1", features = ["no-entrypoint"] }
 spl-type-length-value = "0.5.0"
 kaptn-macros = { version = "0.1.3", path = "macros" }
+kaptn-attribute-transfer-hook = { version = "0.1.0", path = "attribute/transfer_hook" }
+kaptn-derive-extrametas = { version = "0.1.0", path = "derive/extrametas" }
 
 [lib]
 crate-type = ["cdylib", "lib"]
