+++ lang/attribute/meta/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "kaptn-attribute-meta"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[lib]
+proc-macro = true
+
+[dependencies]
+syn = { version = "1", features = ["full", "extra-traits", "parsing"] }
+quote = "1"
+proc-macro2 = "1"
