+++ lang/derive/extrametas/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "kaptn-derive-extrametas"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[lib]
+proc-macro = true
+
+[dependencies]
+syn = { version = "1", features = ["full", "extra-traits", "parsing"] }
+quote = "1"
+proc-macro2 = "1"
+kaptn-attribute-meta = { path = "../../attribute/meta" }
